🗺️ Atlas: [Extract Native Handlers]
Extracted `bootstrap_stdlib` logic and standard library native implementations into a dedicated `crates/duke-interpreter/src/natives.rs` module. Re-exported `bootstrap_stdlib` from `lib.rs` to maintain the public API while significantly reducing file bloat.
