+++ crates/duke-interpreter/src/lib.rs
@@ -4,8 +4,10 @@
 //! float, and double arithmetic, control flow, and local variables.  Heap
 //! allocation, field access, and method invocation are not yet implemented.
 
+pub use natives::bootstrap_stdlib;
 /// Core execution context types for methods and classes.
 pub mod context;
+pub mod natives;
 /// Repositories for loaded classes and registered native methods.
 pub mod registry;
 mod threading;
@@ -39,5304 +41,6 @@ use duke_runtime::{Frame, Slot, VmError, VmResult};
 /// registers native `println` handlers for `(Ljava/lang/String;)V`, `(I)V`,
 /// and `()V`.
 #[allow(clippy::too_many_lines)]
-pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) {
-    // Allocate a PrintStream object on the heap.
-    let ps_ref = heap.allocate("java/io/PrintStream".to_string(), 0);
-
-    // Create java/lang/System ClassContext with a single static field `out`.
-    let system_ctx = ClassContext {
-        class_name: "java/lang/System".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "out".to_string(),
-            descriptor: "Ljava/io/PrintStream;".to_string(),
-            is_static: true,
-        }],
-        static_fields: vec![Slot::Reference(Some(ps_ref))],
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(system_ctx);
-
-    // Create java/io/PrintStream ClassContext (empty — all methods are native).
-    let ps_ctx = ClassContext {
-        class_name: "java/io/PrintStream".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(ps_ctx);
-
-    // Register native println handlers.
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(Ljava/lang/String;)V",
-        native_println_string,
-    );
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "println", "(I)V", native_println_int);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "println", "()V", native_println_void);
-
-    let file_ctx = ClassContext {
-        class_name: "java/io/File".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "path".to_string(),
-            descriptor: "Ljava/lang/String;".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(file_ctx);
-    registry.natives_mut().register(
-        "java/io/File",
-        "<init>",
-        "(Ljava/lang/String;)V",
-        native_file_init,
-    );
-    registry
-        .natives_mut()
-        .register("java/io/File", "exists", "()Z", native_file_exists);
-    registry
-        .natives_mut()
-        .register("java/io/File", "isFile", "()Z", native_file_is_file);
-    registry.natives_mut().register(
-        "java/io/File",
-        "isDirectory",
-        "()Z",
-        native_file_is_directory,
-    );
-
-    let io_exception_ctx = ClassContext {
-        class_name: "java/io/IOException".to_string(),
-        super_class: Some("java/lang/Exception".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(io_exception_ctx);
-
-    let file_not_found_ctx = ClassContext {
-        class_name: "java/io/FileNotFoundException".to_string(),
-        super_class: Some("java/io/IOException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(file_not_found_ctx);
-
-    let socket_exception_ctx = ClassContext {
-        class_name: "java/net/SocketException".to_string(),
-        super_class: Some("java/io/IOException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(socket_exception_ctx);
-
-    let bind_exception_ctx = ClassContext {
-        class_name: "java/net/BindException".to_string(),
-        super_class: Some("java/net/SocketException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(bind_exception_ctx);
-
-    let connect_exception_ctx = ClassContext {
-        class_name: "java/net/ConnectException".to_string(),
-        super_class: Some("java/net/SocketException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(connect_exception_ctx);
-
-    let input_stream_ctx = ClassContext {
-        class_name: "java/io/InputStream".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(input_stream_ctx);
-
-    let output_stream_ctx = ClassContext {
-        class_name: "java/io/OutputStream".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(output_stream_ctx);
-
-    let file_input_stream_ctx = ClassContext {
-        class_name: "java/io/FileInputStream".to_string(),
-        super_class: Some("java/io/InputStream".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "fd".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(file_input_stream_ctx);
-    registry.natives_mut().register(
-        "java/io/FileInputStream",
-        "<init>",
-        "(Ljava/lang/String;)V",
-        native_file_input_stream_init,
-    );
-    registry.natives_mut().register(
-        "java/io/FileInputStream",
-        "read",
-        "()I",
-        native_file_input_stream_read,
-    );
-    registry.natives_mut().register(
-        "java/io/FileInputStream",
-        "read",
-        "([B)I",
-        native_file_input_stream_read_bytes,
-    );
-    registry.natives_mut().register(
-        "java/io/FileInputStream",
-        "close",
-        "()V",
-        native_file_input_stream_close,
-    );
-
-    let file_output_stream_ctx = ClassContext {
-        class_name: "java/io/FileOutputStream".to_string(),
-        super_class: Some("java/io/OutputStream".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "fd".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(file_output_stream_ctx);
-    registry.natives_mut().register(
-        "java/io/FileOutputStream",
-        "<init>",
-        "(Ljava/lang/String;)V",
-        native_file_output_stream_init,
-    );
-    registry.natives_mut().register(
-        "java/io/FileOutputStream",
-        "write",
-        "(I)V",
-        native_file_output_stream_write,
-    );
-    registry.natives_mut().register(
-        "java/io/FileOutputStream",
-        "write",
-        "([B)V",
-        native_file_output_stream_write_bytes,
-    );
-    registry.natives_mut().register(
-        "java/io/FileOutputStream",
-        "close",
-        "()V",
-        native_file_output_stream_close,
-    );
-
-    let server_socket_ctx = ClassContext {
-        class_name: "java/net/ServerSocket".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "fd".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "port".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 2,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(server_socket_ctx);
-    registry.natives_mut().register(
-        "java/net/ServerSocket",
-        "<init>",
-        "(I)V",
-        native_server_socket_init,
-    );
-    registry.natives_mut().register(
-        "java/net/ServerSocket",
-        "accept",
-        "()Ljava/net/Socket;",
-        native_server_socket_accept,
-    );
-    registry.natives_mut().register(
-        "java/net/ServerSocket",
-        "getLocalPort",
-        "()I",
-        native_server_socket_get_local_port,
-    );
-    registry.natives_mut().register(
-        "java/net/ServerSocket",
-        "close",
-        "()V",
-        native_server_socket_close,
-    );
-
-    let socket_ctx = ClassContext {
-        class_name: "java/net/Socket".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "fdRead".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "fdWrite".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 2,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(socket_ctx);
-    registry.natives_mut().register(
-        "java/net/Socket",
-        "<init>",
-        "(Ljava/lang/String;I)V",
-        native_socket_init,
-    );
-    registry.natives_mut().register(
-        "java/net/Socket",
-        "getInputStream",
-        "()Ljava/io/InputStream;",
-        native_socket_get_input_stream,
-    );
-    registry.natives_mut().register(
-        "java/net/Socket",
-        "getOutputStream",
-        "()Ljava/io/OutputStream;",
-        native_socket_get_output_stream,
-    );
-    registry
-        .natives_mut()
-        .register("java/net/Socket", "close", "()V", native_socket_close);
-
-    let socket_input_stream_ctx = ClassContext {
-        class_name: "duke/net/SocketInputStream".to_string(),
-        super_class: Some("java/io/InputStream".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "fd".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(socket_input_stream_ctx);
-    registry.natives_mut().register(
-        "duke/net/SocketInputStream",
-        "read",
-        "()I",
-        native_file_input_stream_read,
-    );
-    registry.natives_mut().register(
-        "duke/net/SocketInputStream",
-        "read",
-        "([B)I",
-        native_file_input_stream_read_bytes,
-    );
-    registry.natives_mut().register(
-        "duke/net/SocketInputStream",
-        "close",
-        "()V",
-        native_file_input_stream_close,
-    );
-
-    let socket_output_stream_ctx = ClassContext {
-        class_name: "duke/net/SocketOutputStream".to_string(),
-        super_class: Some("java/io/OutputStream".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "fd".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(socket_output_stream_ctx);
-    registry.natives_mut().register(
-        "duke/net/SocketOutputStream",
-        "write",
-        "(I)V",
-        native_file_output_stream_write,
-    );
-    registry.natives_mut().register(
-        "duke/net/SocketOutputStream",
-        "write",
-        "([B)V",
-        native_file_output_stream_write_bytes,
-    );
-    registry.natives_mut().register(
-        "duke/net/SocketOutputStream",
-        "close",
-        "()V",
-        native_file_output_stream_close,
-    );
-
-    // Register java/lang/String ClassContext (empty — instance methods are native).
-    let string_ctx = ClassContext {
-        class_name: "java/lang/String".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: vec![
-            "java/lang/Comparable".to_string(),
-            "java/io/Serializable".to_string(),
-            "java/lang/CharSequence".to_string(),
-        ],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(string_ctx);
-
-    // String instance methods
-    registry
-        .natives_mut()
-        .register("java/lang/String", "length", "()I", native_string_length);
-    registry.natives_mut().register(
-        "java/lang/String",
-        "equals",
-        "(Ljava/lang/Object;)Z",
-        native_string_equals,
-    );
-    registry
-        .natives_mut()
-        .register("java/lang/String", "charAt", "(I)C", native_string_char_at);
-    registry.natives_mut().register(
-        "java/lang/String",
-        "substring",
-        "(I)Ljava/lang/String;",
-        native_string_substring,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "substring",
-        "(II)Ljava/lang/String;",
-        native_string_substring_range,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "indexOf",
-        "(Ljava/lang/String;)I",
-        native_string_indexof,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "contains",
-        "(Ljava/lang/CharSequence;)Z",
-        native_string_contains,
-    );
-    registry
-        .natives_mut()
-        .register("java/lang/String", "isEmpty", "()Z", native_string_isempty);
-    registry.natives_mut().register(
-        "java/lang/String",
-        "compareTo",
-        "(Ljava/lang/String;)I",
-        native_string_compareto,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "compareTo",
-        "(Ljava/lang/Object;)I",
-        native_string_compareto_object,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "startsWith",
-        "(Ljava/lang/String;)Z",
-        native_string_startswith,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "endsWith",
-        "(Ljava/lang/String;)Z",
-        native_string_endswith,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "trim",
-        "()Ljava/lang/String;",
-        native_string_trim,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "toCharArray",
-        "()[C",
-        native_string_tochararray,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "toUpperCase",
-        "()Ljava/lang/String;",
-        native_string_touppercase,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "toLowerCase",
-        "()Ljava/lang/String;",
-        native_string_tolowercase,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "replace",
-        "(CC)Ljava/lang/String;",
-        native_string_replace_char,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "replace",
-        "(Ljava/lang/CharSequence;Ljava/lang/CharSequence;)Ljava/lang/String;",
-        native_string_replace_charsequence,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "split",
-        "(Ljava/lang/String;)[Ljava/lang/String;",
-        native_string_split,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "hashCode",
-        "()I",
-        native_string_hashcode,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "toString",
-        "()Ljava/lang/String;",
-        native_string_tostring,
-    );
-
-    // String static methods
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(I)Ljava/lang/String;",
-        native_string_value_of_int,
-    );
-
-    // PrintStream.print (no newline)
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "print",
-        "(Ljava/lang/String;)V",
-        native_print_string,
-    );
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(I)V", native_print_int);
-
-    // println overloads
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(J)V",
-        native_println_long,
-    );
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(F)V",
-        native_println_float,
-    );
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(D)V",
-        native_println_double,
-    );
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(Z)V",
-        native_println_boolean,
-    );
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(C)V",
-        native_println_char,
-    );
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "println",
-        "(Ljava/lang/Object;)V",
-        native_println_object,
-    );
-
-    // print overloads
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(J)V", native_print_long);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(F)V", native_print_float);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(D)V", native_print_double);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(Z)V", native_print_boolean);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(C)V", native_print_char);
-    registry.natives_mut().register(
-        "java/io/PrintStream",
-        "print",
-        "(Ljava/lang/Object;)V",
-        native_print_object,
-    );
-
-    // System.exit (static)
-    registry
-        .natives_mut()
-        .register("java/lang/System", "exit", "(I)V", native_system_exit);
-
-    // Register synthetic exception hierarchy so is_assignable_from can walk it.
-    // java/lang/Object (root — no super)
-    let object_ctx = ClassContext {
-        class_name: "java/lang/Object".to_string(),
-        super_class: None,
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(object_ctx);
-
-    // Object instance methods
-    registry.natives_mut().register(
-        "java/lang/Object",
-        "hashCode",
-        "()I",
-        native_object_hashcode,
-    );
-    registry.natives_mut().register(
-        "java/lang/Object",
-        "toString",
-        "()Ljava/lang/String;",
-        native_object_tostring,
-    );
-    registry.natives_mut().register(
-        "java/lang/Object",
-        "clone",
-        "()Ljava/lang/Object;",
-        native_object_clone,
-    );
-
-    // java/lang/Class — lightweight stub for class literals
-    let class_ctx = ClassContext {
-        class_name: "java/lang/Class".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(class_ctx);
-    registry.natives_mut().register_callback(
-        "java/lang/Class",
-        "forName",
-        "(Ljava/lang/String;)Ljava/lang/Class;",
-        native_class_for_name,
-    );
-    registry.natives_mut().register(
-        "java/lang/Class",
-        "getName",
-        "()Ljava/lang/String;",
-        native_class_get_name,
-    );
-    registry.natives_mut().register_callback(
-        "java/lang/Class",
-        "getDeclaredMethods",
-        "()[Ljava/lang/reflect/Method;",
-        native_class_get_declared_methods,
-    );
-    registry.natives_mut().register_callback(
-        "java/lang/Class",
-        "getDeclaredFields",
-        "()[Ljava/lang/reflect/Field;",
-        native_class_get_declared_fields,
-    );
-
-    let reflect_method_ctx = ClassContext {
-        class_name: "java/lang/reflect/Method".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "declaringClass".to_string(),
-                descriptor: "Ljava/lang/Class;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "name".to_string(),
-                descriptor: "Ljava/lang/String;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "descriptor".to_string(),
-                descriptor: "Ljava/lang/String;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "publicFlag".to_string(),
-                descriptor: "Z".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "staticFlag".to_string(),
-                descriptor: "Z".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 5,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(reflect_method_ctx);
-    registry.natives_mut().register(
-        "java/lang/reflect/Method",
-        "getName",
-        "()Ljava/lang/String;",
-        native_reflect_method_get_name,
-    );
-    registry.natives_mut().register_callback(
-        "java/lang/reflect/Method",
-        "invoke",
-        "(Ljava/lang/Object;[Ljava/lang/Object;)Ljava/lang/Object;",
-        native_reflect_method_invoke,
-    );
-
-    let reflect_field_ctx = ClassContext {
-        class_name: "java/lang/reflect/Field".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "declaringClass".to_string(),
-                descriptor: "Ljava/lang/Class;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "name".to_string(),
-                descriptor: "Ljava/lang/String;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "descriptor".to_string(),
-                descriptor: "Ljava/lang/String;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "publicFlag".to_string(),
-                descriptor: "Z".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "staticFlag".to_string(),
-                descriptor: "Z".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 5,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(reflect_field_ctx);
-    registry.natives_mut().register(
-        "java/lang/reflect/Field",
-        "getName",
-        "()Ljava/lang/String;",
-        native_reflect_field_get_name,
-    );
-
-    let runnable_ctx = ClassContext {
-        class_name: "java/lang/Runnable".to_string(),
-        super_class: None,
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(runnable_ctx);
-
-    let thread_ctx = ClassContext {
-        class_name: "java/lang/Thread".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "target".to_string(),
-                descriptor: "Ljava/lang/Runnable;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "threadId".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 2,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(thread_ctx);
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "<init>", "()V", native_thread_init);
-    registry.natives_mut().register(
-        "java/lang/Thread",
-        "<init>",
-        "(Ljava/lang/Runnable;)V",
-        native_thread_init_runnable,
-    );
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "start", "()V", native_thread_start);
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "join", "()V", native_thread_join);
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "sleep", "(J)V", native_thread_sleep);
-
-    // java/lang/Throwable extends Object
-    let throwable_ctx = ClassContext {
-        class_name: "java/lang/Throwable".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(throwable_ctx);
-    registry.natives_mut().register(
-        "java/lang/Throwable",
-        "addSuppressed",
-        "(Ljava/lang/Throwable;)V",
-        native_throwable_add_suppressed,
-    );
-
-    // java/lang/Exception extends Throwable
-    let exception_ctx = ClassContext {
-        class_name: "java/lang/Exception".to_string(),
-        super_class: Some("java/lang/Throwable".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(exception_ctx);
-
-    // java/lang/RuntimeException extends Exception
-    let rte_ctx = ClassContext {
-        class_name: "java/lang/RuntimeException".to_string(),
-        super_class: Some("java/lang/Exception".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(rte_ctx);
-
-    let illegal_argument_ctx = ClassContext {
-        class_name: "java/lang/IllegalArgumentException".to_string(),
-        super_class: Some("java/lang/RuntimeException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(illegal_argument_ctx);
-
-    let reflective_operation_ctx = ClassContext {
-        class_name: "java/lang/ReflectiveOperationException".to_string(),
-        super_class: Some("java/lang/Exception".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(reflective_operation_ctx);
-
-    let class_not_found_ctx = ClassContext {
-        class_name: "java/lang/ClassNotFoundException".to_string(),
-        super_class: Some("java/lang/ReflectiveOperationException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(class_not_found_ctx);
-
-    let illegal_access_ctx = ClassContext {
-        class_name: "java/lang/IllegalAccessException".to_string(),
-        super_class: Some("java/lang/ReflectiveOperationException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(illegal_access_ctx);
-
-    let invocation_target_ctx = ClassContext {
-        class_name: "java/lang/reflect/InvocationTargetException".to_string(),
-        super_class: Some("java/lang/ReflectiveOperationException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(invocation_target_ctx);
-
-    // java/lang/AutoCloseable — marker interface for try-with-resources.
-    // Registered so is_assignable_from correctly handles queries like
-    // "does TryWithResources$Res implement AutoCloseable?" without
-    // erroring on an unknown class.
-    let autocloseable_ctx = ClassContext {
-        class_name: "java/lang/AutoCloseable".to_string(),
-        super_class: None, // interface — no super class
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(autocloseable_ctx);
-
-    // java/lang/Enum — abstract superclass for all enums.
-    // Fields: name (String) at index 0, ordinal (int) at index 1.
-    let enum_ctx = ClassContext {
-        class_name: "java/lang/Enum".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "name".to_string(),
-                descriptor: "Ljava/lang/String;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "ordinal".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 2,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(enum_ctx);
-    registry.natives_mut().register(
-        "java/lang/Enum",
-        "<init>",
-        "(Ljava/lang/String;I)V",
-        native_enum_init,
-    );
-    registry
-        .natives_mut()
-        .register("java/lang/Enum", "ordinal", "()I", native_enum_ordinal);
-    registry.natives_mut().register(
-        "java/lang/Enum",
-        "name",
-        "()Ljava/lang/String;",
-        native_enum_name,
-    );
-    registry.natives_mut().register(
-        "java/lang/Enum",
-        "valueOf",
-        "(Ljava/lang/Class;Ljava/lang/String;)Ljava/lang/Enum;",
-        native_enum_valueof,
-    );
-
-    // java/lang/Integer — boxed int with value field + numeric constants
-    let integer_ctx = ClassContext {
-        class_name: "java/lang/Integer".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "value".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "MAX_VALUE".to_string(),
-                descriptor: "I".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "MIN_VALUE".to_string(),
-                descriptor: "I".to_string(),
-                is_static: true,
-            },
-        ],
-        static_fields: vec![Slot::Int(i32::MAX), Slot::Int(i32::MIN)],
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/Comparable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(integer_ctx);
-    registry.natives_mut().register(
-        "java/lang/Integer",
-        "parseInt",
-        "(Ljava/lang/String;)I",
-        native_integer_parseint,
-    );
-    registry.natives_mut().register(
-        "java/lang/Integer",
-        "valueOf",
-        "(I)Ljava/lang/Integer;",
-        native_integer_valueof,
-    );
-    registry.natives_mut().register(
-        "java/lang/Integer",
-        "intValue",
-        "()I",
-        native_integer_intvalue,
-    );
-    registry.natives_mut().register(
-        "java/lang/Integer",
-        "toString",
-        "(I)Ljava/lang/String;",
-        native_integer_tostring_static,
-    );
-    registry.natives_mut().register(
-        "java/lang/Integer",
-        "compareTo",
-        "(Ljava/lang/Object;)I",
-        native_integer_compareto,
-    );
-
-    // java/lang/Long — boxed long with value field + numeric constants
-    let long_ctx = ClassContext {
-        class_name: "java/lang/Long".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "value".to_string(),
-                descriptor: "J".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "MAX_VALUE".to_string(),
-                descriptor: "J".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "MIN_VALUE".to_string(),
-                descriptor: "J".to_string(),
-                is_static: true,
-            },
-        ],
-        static_fields: vec![Slot::Long(i64::MAX), Slot::Long(i64::MIN)],
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/Comparable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(long_ctx);
-    registry.natives_mut().register(
-        "java/lang/Long",
-        "parseLong",
-        "(Ljava/lang/String;)J",
-        native_long_parselong,
-    );
-    registry.natives_mut().register(
-        "java/lang/Long",
-        "valueOf",
-        "(J)Ljava/lang/Long;",
-        native_long_valueof,
-    );
-    registry
-        .natives_mut()
-        .register("java/lang/Long", "longValue", "()J", native_long_longvalue);
-    registry.natives_mut().register(
-        "java/lang/Long",
-        "toString",
-        "(J)Ljava/lang/String;",
-        native_long_tostring_static,
-    );
-    registry.natives_mut().register(
-        "java/lang/Long",
-        "compareTo",
-        "(Ljava/lang/Object;)I",
-        native_long_compareto,
-    );
-
-    // java/lang/Double — boxed double with value field + numeric constants
-    let double_ctx = ClassContext {
-        class_name: "java/lang/Double".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "value".to_string(),
-                descriptor: "D".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "MAX_VALUE".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "MIN_VALUE".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "NaN".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "POSITIVE_INFINITY".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "NEGATIVE_INFINITY".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-        ],
-        static_fields: vec![
-            Slot::Double(f64::MAX),
-            Slot::Double(5e-324_f64),
-            Slot::Double(f64::NAN),
-            Slot::Double(f64::INFINITY),
-            Slot::Double(f64::NEG_INFINITY),
-        ],
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/Comparable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(double_ctx);
-    registry.natives_mut().register(
-        "java/lang/Double",
-        "parseDouble",
-        "(Ljava/lang/String;)D",
-        native_double_parsedouble,
-    );
-    registry.natives_mut().register(
-        "java/lang/Double",
-        "valueOf",
-        "(D)Ljava/lang/Double;",
-        native_double_valueof,
-    );
-    registry.natives_mut().register(
-        "java/lang/Double",
-        "doubleValue",
-        "()D",
-        native_double_doublevalue,
-    );
-    registry
-        .natives_mut()
-        .register("java/lang/Double", "isNaN", "(D)Z", native_double_isnan);
-    registry.natives_mut().register(
-        "java/lang/Double",
-        "compareTo",
-        "(Ljava/lang/Object;)I",
-        native_double_compareto,
-    );
-
-    // java/lang/Float — boxed float with value field
-    let float_ctx = ClassContext {
-        class_name: "java/lang/Float".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "value".to_string(),
-            descriptor: "F".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/Comparable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(float_ctx);
-    registry.natives_mut().register(
-        "java/lang/Float",
-        "parseFloat",
-        "(Ljava/lang/String;)F",
-        native_float_parsefloat,
-    );
-
-    // java/lang/Boolean — static utility
-    let boolean_ctx = ClassContext {
-        class_name: "java/lang/Boolean".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: vec!["java/lang/Comparable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(boolean_ctx);
-    registry.natives_mut().register(
-        "java/lang/Boolean",
-        "parseBoolean",
-        "(Ljava/lang/String;)Z",
-        native_boolean_parseboolean,
-    );
-
-    // java/lang/Math — static math utilities with PI and E constants
-    let math_ctx = ClassContext {
-        class_name: "java/lang/Math".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "PI".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-            FieldEntry {
-                name: "E".to_string(),
-                descriptor: "D".to_string(),
-                is_static: true,
-            },
-        ],
-        static_fields: vec![
-            Slot::Double(std::f64::consts::PI),
-            Slot::Double(std::f64::consts::E),
-        ],
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(math_ctx);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "max", "(II)I", native_math_max_int);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "min", "(II)I", native_math_min_int);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "abs", "(I)I", native_math_abs_int);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "sqrt", "(D)D", native_math_sqrt);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "pow", "(DD)D", native_math_pow);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "floor", "(D)D", native_math_floor);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "ceil", "(D)D", native_math_ceil);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "round", "(D)J", native_math_round_double);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "abs", "(J)J", native_math_abs_long);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "abs", "(D)D", native_math_abs_double);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "max", "(JJ)J", native_math_max_long);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "min", "(JJ)J", native_math_min_long);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "max", "(DD)D", native_math_max_double);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "min", "(DD)D", native_math_min_double);
-
-    // String.valueOf overloads (int already registered above)
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(J)Ljava/lang/String;",
-        native_string_value_of_long,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(D)Ljava/lang/String;",
-        native_string_value_of_double,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(F)Ljava/lang/String;",
-        native_string_value_of_float,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(Z)Ljava/lang/String;",
-        native_string_value_of_boolean,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(C)Ljava/lang/String;",
-        native_string_value_of_char,
-    );
-    registry.natives_mut().register(
-        "java/lang/String",
-        "valueOf",
-        "(Ljava/lang/Object;)Ljava/lang/String;",
-        native_string_value_of_object,
-    );
-
-    // String.concat
-    registry.natives_mut().register(
-        "java/lang/String",
-        "concat",
-        "(Ljava/lang/String;)Ljava/lang/String;",
-        native_string_concat,
-    );
-
-    // String.format(String, Object[]) -> String
-    registry.natives_mut().register(
-        "java/lang/String",
-        "format",
-        "(Ljava/lang/String;[Ljava/lang/Object;)Ljava/lang/String;",
-        native_string_format,
-    );
-
-    // java/lang/StringBuilder — mutable string buffer
-    let sb_ctx = ClassContext {
-        class_name: "java/lang/StringBuilder".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: vec!["java/lang/CharSequence".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(sb_ctx);
-
-    // StringBuilder.<init>()V
-    registry
-        .natives_mut()
-        .register("java/lang/StringBuilder", "<init>", "()V", native_sb_init);
-    // StringBuilder.<init>(String)V
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "<init>",
-        "(Ljava/lang/String;)V",
-        native_sb_init_string,
-    );
-    // StringBuilder.append(String)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(Ljava/lang/String;)Ljava/lang/StringBuilder;",
-        native_sb_append_string,
-    );
-    // StringBuilder.append(int)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(I)Ljava/lang/StringBuilder;",
-        native_sb_append_int,
-    );
-    // StringBuilder.append(long)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(J)Ljava/lang/StringBuilder;",
-        native_sb_append_long,
-    );
-    // StringBuilder.append(double)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(D)Ljava/lang/StringBuilder;",
-        native_sb_append_double,
-    );
-    // StringBuilder.append(float)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(F)Ljava/lang/StringBuilder;",
-        native_sb_append_float,
-    );
-    // StringBuilder.append(boolean)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(Z)Ljava/lang/StringBuilder;",
-        native_sb_append_boolean,
-    );
-    // StringBuilder.append(char)
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "append",
-        "(C)Ljava/lang/StringBuilder;",
-        native_sb_append_char,
-    );
-    // StringBuilder.toString()
-    registry.natives_mut().register(
-        "java/lang/StringBuilder",
-        "toString",
-        "()Ljava/lang/String;",
-        native_sb_tostring,
-    );
-    // StringBuilder.length()
-    registry
-        .natives_mut()
-        .register("java/lang/StringBuilder", "length", "()I", native_sb_length);
-
-    // java/lang/Character — static character utilities + boxed char
-    let character_ctx = ClassContext {
-        class_name: "java/lang/Character".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "value".to_string(),
-            descriptor: "C".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/Comparable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(character_ctx);
-
-    // Character.isDigit(C)Z
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "isDigit",
-        "(C)Z",
-        native_char_is_digit,
-    );
-    // Character.isLetter(C)Z
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "isLetter",
-        "(C)Z",
-        native_char_is_letter,
-    );
-    // Character.isWhitespace(C)Z
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "isWhitespace",
-        "(C)Z",
-        native_char_is_whitespace,
-    );
-    // Character.isUpperCase(C)Z
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "isUpperCase",
-        "(C)Z",
-        native_char_is_uppercase,
-    );
-    // Character.isLowerCase(C)Z
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "isLowerCase",
-        "(C)Z",
-        native_char_is_lowercase,
-    );
-    // Character.toUpperCase(C)C
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "toUpperCase",
-        "(C)C",
-        native_char_to_uppercase,
-    );
-    // Character.toLowerCase(C)C
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "toLowerCase",
-        "(C)C",
-        native_char_to_lowercase,
-    );
-    // Character.isLetterOrDigit(C)Z
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "isLetterOrDigit",
-        "(C)Z",
-        native_char_is_letter_or_digit,
-    );
-    // Character.valueOf(C)Ljava/lang/Character;
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "valueOf",
-        "(C)Ljava/lang/Character;",
-        native_char_valueof,
-    );
-    // Character.charValue()C
-    registry.natives_mut().register(
-        "java/lang/Character",
-        "charValue",
-        "()C",
-        native_char_charvalue,
-    );
-
-    // java/util/ArrayList — dynamic list backed by growable fields
-    // fields[0] = size (Int), fields[1..] = elements (pushed dynamically)
-    let arraylist_ctx = ClassContext {
-        class_name: "java/util/ArrayList".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "size".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec![
-            "java/util/List".to_string(),
-            "java/util/Collection".to_string(),
-            "java/lang/Iterable".to_string(),
-        ],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(arraylist_ctx);
-    registry.natives_mut().register(
-        "java/util/ArrayList",
-        "<init>",
-        "()V",
-        native_arraylist_init,
-    );
-    registry.natives_mut().register(
-        "java/util/ArrayList",
-        "add",
-        "(Ljava/lang/Object;)Z",
-        native_arraylist_add,
-    );
-    registry.natives_mut().register(
-        "java/util/ArrayList",
-        "get",
-        "(I)Ljava/lang/Object;",
-        native_arraylist_get,
-    );
-    registry
-        .natives_mut()
-        .register("java/util/ArrayList", "size", "()I", native_arraylist_size);
-    registry.natives_mut().register(
-        "java/util/ArrayList",
-        "iterator",
-        "()Ljava/util/Iterator;",
-        native_arraylist_iterator,
-    );
-    registry.natives_mut().register_callback(
-        "java/util/ArrayList",
-        "sort",
-        "(Ljava/util/Comparator;)V",
-        array_list_sort,
-    );
-
-    // duke/util/ArrayListIterator — internal iterator for ArrayList
-    // fields[0] = ArrayList reference, fields[1] = current index (Int)
-    let iter_ctx = ClassContext {
-        class_name: "duke/util/ArrayListIterator".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "list".to_string(),
-                descriptor: "Ljava/util/ArrayList;".to_string(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "index".to_string(),
-                descriptor: "I".to_string(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 2,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(iter_ctx);
-    registry.natives_mut().register(
-        "duke/util/ArrayListIterator",
-        "<init>",
-        "()V",
-        native_arraylist_iter_init,
-    );
-    registry.natives_mut().register(
-        "duke/util/ArrayListIterator",
-        "hasNext",
-        "()Z",
-        native_arraylist_iter_hasnext,
-    );
-    registry.natives_mut().register(
-        "duke/util/ArrayListIterator",
-        "next",
-        "()Ljava/lang/Object;",
-        native_arraylist_iter_next,
-    );
-
-    // java/util/Arrays — static array utilities
-    let arrays_ctx = ClassContext {
-        class_name: "java/util/Arrays".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(arrays_ctx);
-    registry
-        .natives_mut()
-        .register("java/util/Arrays", "fill", "([II)V", native_arrays_fill_int);
-    registry.natives_mut().register(
-        "java/util/Arrays",
-        "fill",
-        "([Ljava/lang/Object;Ljava/lang/Object;)V",
-        native_arrays_fill_object,
-    );
-    registry.natives_mut().register(
-        "java/util/Arrays",
-        "copyOf",
-        "([II)[I",
-        native_arrays_copyof_int,
-    );
-    registry.natives_mut().register(
-        "java/util/Arrays",
-        "copyOf",
-        "([Ljava/lang/Object;I)[Ljava/lang/Object;",
-        native_arrays_copyof_object,
-    );
-    registry
-        .natives_mut()
-        .register("java/util/Arrays", "sort", "([I)V", native_arrays_sort_int);
-
-    // java/util/HashMap — hash map backed by flat key/value pair list in fields
-    // fields[0] = Int(size), fields[1]=key0, fields[2]=val0, fields[3]=key1, ...
-    let hashmap_ctx = ClassContext {
-        class_name: "java/util/HashMap".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "size".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(hashmap_ctx);
-    registry
-        .natives_mut()
-        .register("java/util/HashMap", "<init>", "()V", native_hashmap_init);
-    registry.natives_mut().register(
-        "java/util/HashMap",
-        "put",
-        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_put,
-    );
-    registry.natives_mut().register(
-        "java/util/HashMap",
-        "get",
-        "(Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_get,
-    );
-    registry.natives_mut().register(
-        "java/util/HashMap",
-        "containsKey",
-        "(Ljava/lang/Object;)Z",
-        native_hashmap_contains_key,
-    );
-    registry
-        .natives_mut()
-        .register("java/util/HashMap", "size", "()I", native_hashmap_size);
-    registry.natives_mut().register(
-        "java/util/HashMap",
-        "remove",
-        "(Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_remove,
-    );
-    registry.natives_mut().register(
-        "java/util/HashMap",
-        "isEmpty",
-        "()Z",
-        native_hashmap_is_empty,
-    );
-    registry.natives_mut().register(
-        "java/util/HashMap",
-        "getOrDefault",
-        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_get_or_default,
-    );
-
-    // java/util/HashSet — set backed by unique elements in fields
-    // fields[0] = Int(size), fields[1..] = elements (unique, no duplicates)
-    let hashset_ctx = ClassContext {
-        class_name: "java/util/HashSet".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "size".to_string(),
-            descriptor: "I".to_string(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(hashset_ctx);
-    registry
-        .natives_mut()
-        .register("java/util/HashSet", "<init>", "()V", native_hashset_init);
-    registry.natives_mut().register(
-        "java/util/HashSet",
-        "add",
-        "(Ljava/lang/Object;)Z",
-        native_hashset_add,
-    );
-    registry.natives_mut().register(
-        "java/util/HashSet",
-        "contains",
-        "(Ljava/lang/Object;)Z",
-        native_hashset_contains,
-    );
-    registry.natives_mut().register(
-        "java/util/HashSet",
-        "remove",
-        "(Ljava/lang/Object;)Z",
-        native_hashset_remove,
-    );
-    registry
-        .natives_mut()
-        .register("java/util/HashSet", "size", "()I", native_hashset_size);
-    registry.natives_mut().register(
-        "java/util/HashSet",
-        "isEmpty",
-        "()Z",
-        native_hashset_is_empty,
-    );
-
-    // java/util/Collections — static utility class
-    // Only sort(List) is supported; delegates to ArrayList.sort(null).
-    let collections_ctx = ClassContext {
-        class_name: "java/util/Collections".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(collections_ctx);
-    registry.natives_mut().register_callback(
-        "java/util/Collections",
-        "sort",
-        "(Ljava/util/List;)V",
-        native_collections_sort,
-    );
-
-    // ── java.util.zip ──────────────────────────────────────────────────
-
-    let zip_exception_ctx = ClassContext {
-        class_name: "java/util/zip/ZipException".to_string(),
-        super_class: Some("java/io/IOException".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(zip_exception_ctx);
-
-    // ZipEntry: [name: Ref, compressedSize_lo: Int, compressedSize_hi: Int,
-    //            size_lo: Int, size_hi: Int, method: Int]  → 6 fields
-    let zip_entry_ctx = ClassContext {
-        class_name: "java/util/zip/ZipEntry".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![
-            FieldEntry {
-                name: "name".into(),
-                descriptor: "Ljava/lang/String;".into(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "compressedSize_lo".into(),
-                descriptor: "I".into(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "compressedSize_hi".into(),
-                descriptor: "I".into(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "size_lo".into(),
-                descriptor: "I".into(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "size_hi".into(),
-                descriptor: "I".into(),
-                is_static: false,
-            },
-            FieldEntry {
-                name: "method".into(),
-                descriptor: "I".into(),
-                is_static: false,
-            },
-        ],
-        static_fields: Vec::new(),
-        instance_field_count: 6,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(zip_entry_ctx);
-    registry.natives_mut().register(
-        "java/util/zip/ZipEntry",
-        "getName",
-        "()Ljava/lang/String;",
-        native_zip_entry_get_name,
-    );
-    registry.natives_mut().register(
-        "java/util/zip/ZipEntry",
-        "getCompressedSize",
-        "()J",
-        native_zip_entry_get_compressed_size,
-    );
-    registry.natives_mut().register(
-        "java/util/zip/ZipEntry",
-        "getSize",
-        "()J",
-        native_zip_entry_get_size,
-    );
-    registry.natives_mut().register(
-        "java/util/zip/ZipEntry",
-        "getMethod",
-        "()I",
-        native_zip_entry_get_method,
-    );
-
-    // ZipFile: [fd: Int] → 1 field
-    let zip_file_ctx = ClassContext {
-        class_name: "java/util/zip/ZipFile".to_string(),
-        super_class: Some("java/lang/Object".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "fd".into(),
-            descriptor: "I".into(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(zip_file_ctx);
-    registry.natives_mut().register(
-        "java/util/zip/ZipFile",
-        "<init>",
-        "(Ljava/lang/String;)V",
-        native_zip_file_init,
-    );
-    registry.natives_mut().register(
-        "java/util/zip/ZipFile",
-        "getEntry",
-        "(Ljava/lang/String;)Ljava/util/zip/ZipEntry;",
-        native_zip_file_get_entry,
-    );
-    registry.natives_mut().register(
-        "java/util/zip/ZipFile",
-        "getInputStream",
-        "(Ljava/util/zip/ZipEntry;)Ljava/io/InputStream;",
-        native_zip_file_get_input_stream,
-    );
-    registry.natives_mut().register(
-        "java/util/zip/ZipFile",
-        "close",
-        "()V",
-        native_zip_file_close,
-    );
-    registry
-        .natives_mut()
-        .register("java/util/zip/ZipFile", "size", "()I", native_zip_file_size);
-
-    // JarFile extends ZipFile — inherits everything for now.
-    let jar_file_ctx = ClassContext {
-        class_name: "java/util/jar/JarFile".to_string(),
-        super_class: Some("java/util/zip/ZipFile".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: Vec::new(),
-        static_fields: Vec::new(),
-        instance_field_count: 0,
-        interfaces: Vec::new(),
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(jar_file_ctx);
-
-    // ByteBufferInputStream — internal class for reading decompressed ZIP data.
-    let byte_buffer_is_ctx = ClassContext {
-        class_name: "duke/zip/ByteBufferInputStream".to_string(),
-        super_class: Some("java/io/InputStream".to_string()),
-        constant_pool: Vec::new(),
-        methods: Vec::new(),
-        fields: vec![FieldEntry {
-            name: "fd".into(),
-            descriptor: "I".into(),
-            is_static: false,
-        }],
-        static_fields: Vec::new(),
-        instance_field_count: 1,
-        interfaces: vec!["java/lang/AutoCloseable".to_string()],
-        bootstrap_methods: Vec::new(),
-    };
-    registry.register(byte_buffer_is_ctx);
-    registry.natives_mut().register(
-        "duke/zip/ByteBufferInputStream",
-        "read",
-        "()I",
-        native_file_input_stream_read,
-    );
-    registry.natives_mut().register(
-        "duke/zip/ByteBufferInputStream",
-        "read",
-        "([B)I",
-        native_file_input_stream_read_bytes,
-    );
-    registry.natives_mut().register(
-        "duke/zip/ByteBufferInputStream",
-        "close",
-        "()V",
-        native_file_input_stream_close,
-    );
-}
-
-fn native_println_string(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let string_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => {
-            writeln!(out, "null").ok();
-            return Ok(None);
-        }
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(string_ref)?;
-    let text = obj.string_value.as_deref().unwrap_or("null");
-    writeln!(out, "{text}").ok();
-    Ok(None)
-}
-
-fn native_println_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-#[allow(clippy::unnecessary_wraps)] // must match NativeHandler signature
-fn native_println_void(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    writeln!(out).ok();
-    Ok(None)
-}
-
-fn path_from_string_slot(
-    args: &[Slot],
-    idx: usize,
-    heap: &duke_gc::Heap,
-) -> VmResult<std::path::PathBuf> {
-    let path_ref = match args.get(idx) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = heap
-        .get(path_ref)?
-        .string_value
-        .clone()
-        .ok_or(VmError::NullPointerException)?;
-    Ok(std::path::PathBuf::from(path))
-}
-
-fn file_path_from_this(args: &[Slot], heap: &duke_gc::Heap) -> VmResult<std::path::PathBuf> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_ref = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = heap
-        .get(path_ref)?
-        .string_value
-        .clone()
-        .ok_or(VmError::NullPointerException)?;
-    Ok(std::path::PathBuf::from(path))
-}
-
-fn native_file_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let file_obj = heap.get_mut(this_ref)?;
-    let Some(path_field) = file_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *path_field = path_slot;
-    Ok(None)
-}
-
-fn native_file_exists(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let path = file_path_from_this(args, heap)?;
-    Ok(Some(Slot::Int(i32::from(path.exists()))))
-}
-
-fn native_file_is_file(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let path = file_path_from_this(args, heap)?;
-    Ok(Some(Slot::Int(i32::from(path.is_file()))))
-}
-
-fn native_file_is_directory(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let path = file_path_from_this(args, heap)?;
-    Ok(Some(Slot::Int(i32::from(path.is_dir()))))
-}
-
-fn file_stream_id_from_this(args: &[Slot], heap: &duke_gc::Heap) -> VmResult<i32> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => Ok(*id),
-        _ => Err(VmError::JavaException {
-            class_name: "java/io/IOException".to_string(),
-        }),
-    }
-}
-
-fn native_file_input_stream_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = path_from_string_slot(args, 1, heap)?;
-    let file_id = heap.open_host_input_file(&path)?;
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(file_id);
-    Ok(None)
-}
-
-fn native_file_input_stream_read(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    Ok(Some(Slot::Int(heap.read_host_file_byte(file_id)?)))
-}
-
-fn native_file_input_stream_read_bytes(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    let array_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let len = heap.get(array_ref)?.fields.len();
-    if len == 0 {
-        return Ok(Some(Slot::Int(0)));
-    }
-
-    let mut count = 0_usize;
-    for idx in 0..len {
-        let next = heap.read_host_file_byte(file_id)?;
-        if next < 0 {
-            break;
-        }
-        heap.get_mut(array_ref)?.fields[idx] = Slot::Int(next);
-        count += 1;
-    }
-
-    if count == 0 {
-        Ok(Some(Slot::Int(-1)))
-    } else {
-        Ok(Some(Slot::Int(i32::try_from(count).unwrap_or(i32::MAX))))
-    }
-}
-
-fn native_file_input_stream_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let file_id = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".to_string(),
-            });
-        }
-    };
-    heap.close_host_file(file_id);
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(0);
-    Ok(None)
-}
-
-fn native_file_output_stream_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = path_from_string_slot(args, 1, heap)?;
-    let file_id = heap.open_host_output_file(&path)?;
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(file_id);
-    Ok(None)
-}
-
-fn native_file_output_stream_write(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    let value = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    heap.write_host_file_byte(file_id, value)?;
-    Ok(None)
-}
-
-fn native_file_output_stream_write_bytes(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    let array_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let bytes = heap.get(array_ref)?.fields.clone();
-    for byte in bytes {
-        let Slot::Int(value) = byte else {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        };
-        heap.write_host_file_byte(file_id, value)?;
-    }
-    Ok(None)
-}
-
-fn native_file_output_stream_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let file_id = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".to_string(),
-            });
-        }
-    };
-    heap.close_host_file(file_id);
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(0);
-    Ok(None)
-}
-
-// ── Networking natives ────────────────────────────────────────────────────
-
-/// Native: `ServerSocket.<init>(int port)` — binds to 0.0.0.0:{port}.
-fn native_server_socket_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let port = match args.get(1) {
-        Some(Slot::Int(p)) => *p,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let addr = format!("0.0.0.0:{port}");
-    let server_id = heap.bind_server_socket(&addr)?;
-    let actual_port = heap.server_socket_local_port(server_id)?;
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.len() < 2 {
-        return Err(VmError::InvalidRef { address: this_ref });
-    }
-    obj.fields[0] = Slot::Int(server_id);
-    obj.fields[1] = Slot::Int(actual_port);
-    Ok(None)
-}
-
-/// Native: `ServerSocket.accept()` — blocks until a client connects, returns a Socket.
-fn native_server_socket_accept(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let server_fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let (reader_id, writer_id) = heap.accept_connection(server_fd)?;
-    // Allocate a new Socket object with fdRead=reader_id, fdWrite=writer_id
-    let socket_ref = heap.allocate("java/net/Socket".to_string(), 2);
-    heap.get_mut(socket_ref)?.fields[0] = Slot::Int(reader_id);
-    heap.get_mut(socket_ref)?.fields[1] = Slot::Int(writer_id);
-    Ok(Some(Slot::Reference(Some(socket_ref))))
-}
-
-/// Native: `ServerSocket.getLocalPort()` — returns the bound port.
-fn native_server_socket_get_local_port(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.get(1) {
-        Some(Slot::Int(port)) => Ok(Some(Slot::Int(*port))),
-        _ => Err(VmError::JavaException {
-            class_name: "java/io/IOException".into(),
-        }),
-    }
-}
-
-/// Native: `ServerSocket.close()` — closes the OS listener and zeros the fd field.
-fn native_server_socket_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        Some(Slot::Int(_)) => return Ok(None), // already closed — idempotent
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    heap.close_host_file(fd);
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(0);
-    obj.fields[1] = Slot::Int(0); // also zero cached port so getLocalPort() returns 0 after close
-    Ok(None)
-}
-
-/// Native: `Socket.<init>(String host, int port)` — connects to host:port.
-fn native_socket_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let host = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap
-            .get(*r)?
-            .string_value
-            .clone()
-            .ok_or(VmError::NullPointerException)?,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let port = match args.get(2) {
-        Some(Slot::Int(p)) => *p,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let addr = format!("{host}:{port}");
-    let (reader_id, writer_id) = heap.connect_socket(&addr)?;
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.len() < 2 {
-        return Err(VmError::InvalidRef { address: this_ref });
-    }
-    obj.fields[0] = Slot::Int(reader_id);
-    obj.fields[1] = Slot::Int(writer_id);
-    Ok(None)
-}
-
-/// Native: `Socket.getInputStream()` — allocates a `SocketInputStream` wrapping `fdRead`.
-fn native_socket_get_input_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd_read = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let stream_ref = heap.allocate("duke/net/SocketInputStream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(fd_read);
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `Socket.getOutputStream()` — allocates a `SocketOutputStream` wrapping `fdWrite`.
-fn native_socket_get_output_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd_write = match heap.get(this_ref)?.fields.get(1) {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let stream_ref = heap.allocate("duke/net/SocketOutputStream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(fd_write);
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `Socket.close()` — closes both OS handles (fdRead and fdWrite).
-fn native_socket_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd_read = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let fd_write = match heap.get(this_ref)?.fields.get(1) {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    heap.close_host_file(fd_read);
-    heap.close_host_file(fd_write);
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(0);
-    obj.fields[1] = Slot::Int(0);
-    Ok(None)
-}
-
-// ── ZIP / JAR natives ──────────────────────────────────────────────────
-
-/// Native: `ZipFile.<init>(String)` — open and index a ZIP/JAR archive.
-fn native_zip_file_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_str = heap
-        .get(path_ref)?
-        .string_value
-        .as_deref()
-        .ok_or(VmError::NullPointerException)?
-        .to_string();
-    let fd = heap.open_host_zip(std::path::Path::new(&path_str))?;
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(fd);
-    Ok(None)
-}
-
-/// Native: `ZipFile.getEntry(String) -> ZipEntry` — look up an entry by name.
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_zip_file_get_entry(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let name_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let entry_name = heap
-        .get(name_ref)?
-        .string_value
-        .as_deref()
-        .ok_or(VmError::NullPointerException)?
-        .to_string();
-    let info = heap.zip_get_entry_info(fd, &entry_name)?;
-    let Some(info) = info else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    // Allocate a ZipEntry HeapObject with 6 fields.
-    let name_heap_ref = heap.allocate_string(info.name);
-    let entry_ref = heap.allocate("java/util/zip/ZipEntry".to_string(), 6);
-    let entry_obj = heap.get_mut(entry_ref)?;
-    entry_obj.fields[0] = Slot::Reference(Some(name_heap_ref));
-    entry_obj.fields[1] = Slot::Int(info.compressed_size as i32);
-    entry_obj.fields[2] = Slot::Int((info.compressed_size >> 32) as i32);
-    entry_obj.fields[3] = Slot::Int(info.uncompressed_size as i32);
-    entry_obj.fields[4] = Slot::Int((info.uncompressed_size >> 32) as i32);
-    entry_obj.fields[5] = Slot::Int(i32::from(info.compression_method));
-    Ok(Some(Slot::Reference(Some(entry_ref))))
-}
-
-/// Native: `ZipFile.getInputStream(ZipEntry) -> InputStream`
-fn native_zip_file_get_input_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let entry_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    // Get entry name from the ZipEntry object.
-    let name_slot_ref = match heap.get(entry_ref)?.fields.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let entry_name = heap
-        .get(name_slot_ref)?
-        .string_value
-        .as_deref()
-        .ok_or(VmError::NullPointerException)?
-        .to_string();
-    // Decompress the entry and wrap in a ByteBuffer.
-    let data = heap.zip_read_entry(fd, &entry_name)?;
-    let buf_fd = heap.open_host_byte_buffer(data);
-    let is_ref = heap.allocate("duke/zip/ByteBufferInputStream".to_string(), 1);
-    let is_obj = heap.get_mut(is_ref)?;
-    is_obj.fields[0] = Slot::Int(buf_fd);
-    Ok(Some(Slot::Reference(Some(is_ref))))
-}
-
-/// Native: `ZipFile.close()`
-fn native_zip_file_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => return Ok(None),
-    };
-    heap.close_host_file(fd);
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(0);
-    Ok(None)
-}
-
-/// Native: `ZipFile.size() -> int`
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_zip_file_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let count = heap.zip_entry_count(fd)?;
-    Ok(Some(Slot::Int(count as i32)))
-}
-
-/// Native: `ZipEntry.getName() -> String`
-fn native_zip_entry_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(heap.get(this_ref)?.fields[0]))
-}
-
-/// Native: `ZipEntry.getCompressedSize() -> long`
-fn native_zip_entry_get_compressed_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fields = &heap.get(this_ref)?.fields;
-    let lo = match fields[1] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let hi = match fields[2] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let val = (i64::from(hi) << 32) | (i64::from(lo) & 0xFFFF_FFFF);
-    Ok(Some(Slot::Long(val)))
-}
-
-/// Native: `ZipEntry.getSize() -> long`
-fn native_zip_entry_get_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fields = &heap.get(this_ref)?.fields;
-    let lo = match fields[3] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let hi = match fields[4] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let val = (i64::from(hi) << 32) | (i64::from(lo) & 0xFFFF_FFFF);
-    Ok(Some(Slot::Long(val)))
-}
-
-/// Native: `ZipEntry.getMethod() -> int`
-fn native_zip_entry_get_method(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(heap.get(this_ref)?.fields[5]))
-}
-
-/// Native: `String.length()` — returns string length as int.
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_string_length(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    let len = obj.string_value.as_ref().map_or(0, String::len);
-    Ok(Some(Slot::Int(len as i32)))
-}
-
-/// Native: `String.equals(Object)` — compares string content.
-fn native_string_equals(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let other_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Ok(Some(Slot::Int(0))),
-    };
-    // Fetch objects from heap in one go to keep borrows short
-    let this_obj = heap.get(this_ref)?;
-    let other_obj = heap.get(other_ref)?;
-    let this_str = this_obj.string_value.as_deref().unwrap_or_default();
-    let other_str = other_obj.string_value.as_deref().unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(this_str == other_str))))
-}
-
-/// Native: `String.charAt(int)` — returns char at index as int.
-#[allow(clippy::cast_sign_loss)]
-fn native_string_char_at(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let index = match args.get(1) {
-        Some(Slot::Int(i)) => *i,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(this_ref)?;
-    let s = obj.string_value.as_deref().unwrap_or("");
-    let ch = s
-        .chars()
-        .nth(index as usize)
-        .ok_or(VmError::ArrayIndexOutOfBounds {
-            index,
-            length: s.len(),
-        })?;
-    Ok(Some(Slot::Int(ch as i32)))
-}
-
-/// Native: `Object.hashCode()` — returns heap address as hash.
-fn native_object_hashcode(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-        Some(Slot::Reference(Some(r))) => Ok(Some(Slot::Int(*r as i32))),
-        _ => Err(VmError::NullPointerException),
-    }
-}
-
-/// Native: `Object.toString()` — delegates to `heap_object_to_string` so String,
-/// boxed primitives, and opaque objects all produce the correct Java representation.
-fn native_object_tostring(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap_object_to_string(heap.get(this_ref)?, this_ref);
-    let r = heap.allocate_string(s);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Object.clone()` — shallow-copies a heap object.
-fn native_object_clone(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    let cloned_class = obj.class_name.clone();
-    let cloned_fields = obj.fields.clone();
-    let cloned_string = obj.string_value.clone();
-    let new_ref = heap.allocate(cloned_class, 0);
-    let dest = heap.get_mut(new_ref)?;
-    dest.fields = cloned_fields;
-    dest.string_value = cloned_string;
-    Ok(Some(Slot::Reference(Some(new_ref))))
-}
-
-/// `Throwable.addSuppressed(Throwable suppressed)V`
-///
-/// No-op stub. Control flow is handled entirely by the bytecode desugaring —
-/// `addSuppressed` only affects what `getSuppressed()` returns, which is not
-/// yet implemented. Suppressed exception is silently dropped.
-///
-/// Signature: `args[0]` = this (Throwable), `args[1]` = suppressed (Throwable)
-#[allow(clippy::unnecessary_wraps)]
-fn native_throwable_add_suppressed(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _stdout: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    Ok(None)
-}
-
-/// Native: `Enum.<init>(Ljava/lang/String;I)V` — stores name + ordinal.
-/// args: `[this_ref, name_ref, ordinal_int]`
-fn native_enum_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let name_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let ordinal = match args.get(2) {
-        Some(Slot::Int(v)) => *v,
-        _ => 0,
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.len() >= 2 {
-        obj.fields[0] = name_slot;
-        obj.fields[1] = Slot::Int(ordinal);
-    }
-    Ok(None)
-}
-
-/// Native: `Enum.ordinal()I`
-fn native_enum_ordinal(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    match obj.fields.get(1) {
-        Some(Slot::Int(v)) => Ok(Some(Slot::Int(*v))),
-        _ => Ok(Some(Slot::Int(0))),
-    }
-}
-
-/// Native: `Enum.name()Ljava/lang/String;`
-fn native_enum_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    match obj.fields.first() {
-        Some(slot @ Slot::Reference(_)) => Ok(Some(*slot)),
-        _ => Ok(Some(Slot::Reference(None))),
-    }
-}
-
-/// Native: `Enum.valueOf(Ljava/lang/Class;Ljava/lang/String;)Ljava/lang/Enum;`
-/// Searches heap for enum constants of the given class matching the name.
-#[allow(clippy::cast_possible_truncation)]
-fn native_enum_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let class_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let name_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let target_name = heap.get(name_ref)?.string_value.clone().unwrap_or_default();
-    let enum_class_name = heap
-        .get(class_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-
-    let obj_count = heap.len();
-    for i in 0..obj_count {
-        let obj = heap.get(i as u64)?;
-        if obj.class_name == enum_class_name
-            && obj.fields.len() >= 2
-            && let Some(Slot::Reference(Some(name_r))) = obj.fields.first()
-            && let Ok(name_obj) = heap.get(*name_r)
-            && name_obj.string_value.as_deref() == Some(target_name.as_str())
-        {
-            return Ok(Some(Slot::Reference(Some(i as u64))));
-        }
-    }
-
-    Err(VmError::JavaException {
-        class_name: "java/lang/IllegalArgumentException".to_string(),
-    })
-}
-
-// ---------------------------------------------------------------------------
-// Reflection natives
-// ---------------------------------------------------------------------------
-
-fn native_class_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let class_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let internal_name = class_internal_name_from_ref(heap, class_ref)?;
-    let name_ref = heap.allocate_string(internal_name_to_binary_name(&internal_name));
-    Ok(Some(Slot::Reference(Some(name_ref))))
-}
-
-fn native_class_for_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> VmResult<Option<Slot>> {
-    let name_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let binary_name = heap
-        .get(name_ref)?
-        .string_value
-        .clone()
-        .ok_or(VmError::NullPointerException)?;
-    let internal_name = binary_name_to_internal_name(&binary_name);
-    match ops.ensure_loaded(&internal_name) {
-        Ok(()) => {
-            let class_ref = allocate_class_object(heap, &internal_name)?;
-            Ok(Some(Slot::Reference(Some(class_ref))))
-        }
-        Err(VmError::ClassNotFound { .. }) => Err(VmError::JavaException {
-            class_name: "java/lang/ClassNotFoundException".to_string(),
-        }),
-        Err(err) => Err(err),
-    }
-}
-
-fn native_class_get_declared_methods(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> VmResult<Option<Slot>> {
-    let class_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let internal_name = class_internal_name_from_ref(heap, class_ref)?;
-    let reflected = ops.inspect_class(&internal_name)?;
-    let method_refs = reflected
-        .methods
-        .into_iter()
-        .filter(|method| method.name != "<init>" && method.name != "<clinit>")
-        .map(|method| {
-            allocate_reflection_member_object(
-                heap,
-                "java/lang/reflect/Method",
-                &reflected.internal_name,
-                &method.name,
-                &method.descriptor,
-                method.is_public,
-                method.is_static,
-            )
-        })
-        .collect::<VmResult<Vec<_>>>()?;
-    let array_ref = allocate_reference_array(heap, "[Ljava/lang/reflect/Method;", &method_refs)?;
-    Ok(Some(Slot::Reference(Some(array_ref))))
-}
-
-fn native_class_get_declared_fields(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> VmResult<Option<Slot>> {
-    let class_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let internal_name = class_internal_name_from_ref(heap, class_ref)?;
-    let reflected = ops.inspect_class(&internal_name)?;
-    let field_refs = reflected
-        .fields
-        .into_iter()
-        .map(|field| {
-            allocate_reflection_member_object(
-                heap,
-                "java/lang/reflect/Field",
-                &reflected.internal_name,
-                &field.name,
-                &field.descriptor,
-                field.is_public,
-                field.is_static,
-            )
-        })
-        .collect::<VmResult<Vec<_>>>()?;
-    let array_ref = allocate_reference_array(heap, "[Ljava/lang/reflect/Field;", &field_refs)?;
-    Ok(Some(Slot::Reference(Some(array_ref))))
-}
-
-fn native_reflect_method_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let method_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(reflection_member_name_slot(heap, method_ref)?))
-}
-
-fn native_reflect_field_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let field_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(reflection_member_name_slot(heap, field_ref)?))
-}
-
-fn native_reflect_method_invoke(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    output: &mut dyn Write,
-    _control: &mut NativeControl,
-    ops: &mut dyn CallbackOps,
-) -> VmResult<Option<Slot>> {
-    let method_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let target_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let invoke_arg_slots =
-        reflection_array_elements(heap, args.get(2).copied().unwrap_or(Slot::Reference(None)))?;
-    let method = reflected_method_handle(heap, method_ref)?;
-
-    if !method.is_public {
-        return Err(VmError::JavaException {
-            class_name: "java/lang/IllegalAccessException".to_string(),
-        });
-    }
-
-    let invoke_args = build_reflection_invoke_args(
-        heap,
-        target_slot,
-        &method.descriptor,
-        invoke_arg_slots,
-        method.is_static,
-    )?;
-
-    ops.ensure_loaded(&method.declaring_internal_name)?;
-    match ops.invoke(
-        heap,
-        output,
-        &method.declaring_internal_name,
-        &method.method_name,
-        &method.descriptor,
-        invoke_args,
-    ) {
-        Ok(result) => Ok(Some(box_reflection_return_value(
-            heap,
-            descriptor_return_type(&method.descriptor),
-            result,
-        )?)),
-        Err(VmError::JavaException { .. }) => Err(VmError::JavaException {
-            class_name: "java/lang/reflect/InvocationTargetException".to_string(),
-        }),
-        Err(err) => Err(err),
-    }
-}
-
-/// Native: `String.valueOf(int)` — static method, returns string of int.
-fn native_string_value_of_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let s = val.to_string();
-    let r = heap.allocate_string(s);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `PrintStream.print(String)` — no newline.
-fn native_print_string(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let string_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => {
-            write!(out, "null").ok();
-            return Ok(None);
-        }
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(string_ref)?;
-    let text = obj.string_value.as_deref().unwrap_or("null");
-    write!(out, "{text}").ok();
-    Ok(None)
-}
-
-/// Native: `PrintStream.print(int)` — no newline.
-#[allow(clippy::unnecessary_wraps)]
-fn native_print_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// println overloads (long, float, double, boolean, char, object)
-// ---------------------------------------------------------------------------
-
-fn native_println_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_float(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_boolean(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(boolean)",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_char(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(char)",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-#[allow(clippy::cast_possible_wrap)]
-/// Convert a heap object to its Java display string.
-///
-/// Checks `string_value` first (handles String/StringBuilder).
-/// For boxed primitives, extracts the stored value from `fields[0]`.
-/// Falls back to `class_name@hex_ref` for opaque objects.
-fn heap_object_to_string(obj: &duke_gc::HeapObject, obj_ref: u64) -> String {
-    if let Some(s) = &obj.string_value {
-        return s.clone();
-    }
-    match obj.class_name.as_str() {
-        "java/lang/Integer" => {
-            if let Some(Slot::Int(v)) = obj.fields.first() {
-                return v.to_string();
-            }
-        }
-        "java/lang/Long" => {
-            if let Some(Slot::Long(v)) = obj.fields.first() {
-                return v.to_string();
-            }
-        }
-        "java/lang/Double" => {
-            if let Some(Slot::Double(v)) = obj.fields.first() {
-                return format_java_double(*v);
-            }
-        }
-        "java/lang/Float" => {
-            if let Some(Slot::Float(v)) = obj.fields.first() {
-                return format_java_float(*v);
-            }
-        }
-        "java/lang/Boolean" => {
-            return match obj.fields.first() {
-                Some(Slot::Int(v)) if *v != 0 => "true".to_string(),
-                _ => "false".to_string(),
-            };
-        }
-        "java/lang/Character" => {
-            if let Some(Slot::Int(v)) = obj.fields.first()
-                && let Some(c) = char::from_u32((*v).cast_unsigned())
-            {
-                return c.to_string();
-            }
-        }
-        _ => {}
-    }
-    format!("{}@{:x}", obj.class_name, obj_ref)
-}
-
-fn native_println_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.get(1) {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap_object_to_string(heap.get(*r)?, *r);
-            writeln!(out, "{s}").ok();
-        }
-        Some(Slot::Reference(None)) => {
-            writeln!(out, "null").ok();
-        }
-        _ => {
-            writeln!(out, "<unknown>").ok();
-        }
-    }
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// print overloads (long, float, double, boolean, char, object)
-// ---------------------------------------------------------------------------
-
-fn native_print_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_float(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_boolean(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(boolean)",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_char(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(char)",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_print_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.get(1) {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap_object_to_string(heap.get(*r)?, *r);
-            write!(out, "{s}").ok();
-        }
-        Some(Slot::Reference(None)) => {
-            write!(out, "null").ok();
-        }
-        _ => {
-            write!(out, "<unknown>").ok();
-        }
-    }
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// System.exit
-// ---------------------------------------------------------------------------
-
-fn native_system_exit(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let code = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => 1,
-    };
-    Err(VmError::SystemExit { code })
-}
-
-const THREAD_TARGET_SLOT: usize = 0;
-const THREAD_ID_SLOT: usize = 1;
-
-fn native_thread_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let this = heap.get_mut(this_ref)?;
-    this.fields[THREAD_TARGET_SLOT] = Slot::Reference(None);
-    this.fields[THREAD_ID_SLOT] = Slot::Int(-1);
-    Ok(None)
-}
-
-fn native_thread_init_runnable(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let target = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let this = heap.get_mut(this_ref)?;
-    this.fields[THREAD_TARGET_SLOT] = target;
-    this.fields[THREAD_ID_SLOT] = Slot::Int(-1);
-    Ok(None)
-}
-
-fn native_thread_start(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let thread_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    control.request(NativeThreadAction::Start { thread_ref });
-    Ok(None)
-}
-
-fn native_thread_join(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let thread_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let thread = heap.get(thread_ref)?;
-    let Some(Slot::Int(thread_id)) = thread.fields.get(THREAD_ID_SLOT) else {
-        return Ok(None);
-    };
-    if *thread_id >= 0 {
-        control.request(NativeThreadAction::Join {
-            thread_id: *thread_id,
-        });
-    }
-    Ok(None)
-}
-
-fn native_thread_sleep(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let millis = match args.first() {
-        Some(Slot::Long(value)) => *value,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "long",
-                got: "other",
-            });
-        }
-    };
-    let millis = u64::try_from(millis.max(0)).unwrap_or(0);
-    control.request(NativeThreadAction::Sleep(std::time::Duration::from_millis(
-        millis,
-    )));
-    Ok(None)
-}
-
-/// Native: `String.substring(int)` — substring from begin to end.
-#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-fn native_string_substring(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let begin = match args.get(1) {
-        Some(Slot::Int(v)) => *v as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let sub = {
-        let obj = heap.get(this_ref)?;
-        let s = obj.string_value.as_deref().unwrap_or_default();
-        if begin > s.len() {
-            return Err(VmError::ArrayIndexOutOfBounds {
-                index: i32::try_from(begin).unwrap_or(i32::MAX),
-                length: s.len(),
-            });
-        }
-        s.chars().skip(begin).collect::<String>()
-    };
-
-    let r = heap.allocate_string(sub);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.substring(int, int)` — substring from begin to end (exclusive).
-#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-fn native_string_substring_range(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let begin = match args.get(1) {
-        Some(Slot::Int(v)) => *v as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let end = match args.get(2) {
-        Some(Slot::Int(v)) => *v as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let sub = {
-        let obj = heap.get(this_ref)?;
-        let s = obj.string_value.as_deref().unwrap_or_default();
-        if begin > end || end > s.len() {
-            return Err(VmError::ArrayIndexOutOfBounds {
-                index: i32::try_from(end).unwrap_or(i32::MAX),
-                length: s.len(),
-            });
-        }
-        s.chars().skip(begin).take(end - begin).collect::<String>()
-    };
-
-    let r = heap.allocate_string(sub);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.indexOf(String)` — find first occurrence of target.
-fn native_string_indexof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let target_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    // Fetch objects from heap in one go to keep borrows short
-    let this_obj = heap.get(this_ref)?;
-    let target_obj = heap.get(target_ref)?;
-    let s = this_obj.string_value.as_deref().unwrap_or_default();
-    let target = target_obj.string_value.as_deref().unwrap_or_default();
-
-    #[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-    let result = s.find(target).map_or(-1, |i| i as i32);
-    Ok(Some(Slot::Int(result)))
-}
-
-/// Native: `String.contains(CharSequence)` — check if string contains target.
-fn native_string_contains(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let target_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    // Fetch objects from heap in one go to keep borrows short
-    let this_obj = heap.get(this_ref)?;
-    let target_obj = heap.get(target_ref)?;
-    let s = this_obj.string_value.as_deref().unwrap_or_default();
-    let target = target_obj.string_value.as_deref().unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.contains(target)))))
-}
-
-/// Native: `String.isEmpty()` — check if string is empty.
-#[allow(clippy::unnecessary_wraps)]
-fn native_string_isempty(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    let s = obj.string_value.as_deref().unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.is_empty()))))
-}
-
-// Dispatch string constants used by the callback-based sort chain.
-const COMPARE_TO_METHOD: &str = "compareTo";
-const COMPARE_TO_OBJECT_DESC: &str = "(Ljava/lang/Object;)I";
-const SORT_COMPARATOR_DESC: &str = "(Ljava/util/Comparator;)V";
-
-/// Maps a `std::cmp::Ordering` to the Java `compareTo` convention: -1 / 0 / 1.
-///
-/// Used by all boxed-type `compareTo` natives to return a consistent,
-/// sign-correct value without relying on `Ordering`'s internal discriminant.
-#[inline]
-const fn ordering_to_int(o: std::cmp::Ordering) -> i32 {
-    match o {
-        std::cmp::Ordering::Less => -1,
-        std::cmp::Ordering::Equal => 0,
-        std::cmp::Ordering::Greater => 1,
-    }
-}
-
-/// Native: `String.compareTo(String)` — delegates to the Object overload.
-fn native_string_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    native_string_compareto_object(args, heap, out, control)
-}
-
-/// Native: `String.compareTo(Object)` — lexicographic comparison via Object descriptor.
-fn native_string_compareto_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_val = |s: &Slot| -> VmResult<String> {
-        match s {
-            Slot::Reference(Some(r)) => Ok(heap.get(*r)?.string_value.clone().unwrap_or_default()),
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => str_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => str_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(Slot::Int(ordering_to_int(a.as_str().cmp(b.as_str())))))
-}
-
-/// Native: `String.startsWith(String)` — check if string starts with prefix.
-fn native_string_startswith(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let prefix_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let prefix = heap
-        .get(prefix_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.starts_with(&prefix)))))
-}
-
-/// Native: `String.endsWith(String)` — check if string ends with suffix.
-fn native_string_endswith(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let suffix_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let suffix = heap
-        .get(suffix_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.ends_with(&suffix)))))
-}
-
-/// Native: `String.trim()` — remove leading and trailing whitespace.
-fn native_string_trim(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let trimmed = s.trim().to_string();
-    let r = heap.allocate_string(trimmed);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.toCharArray()` — convert string to char array.
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_string_tochararray(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let chars: Vec<char> = s.chars().collect();
-    let arr_ref = heap.allocate("[C".to_string(), chars.len());
-    for (i, &c) in chars.iter().enumerate() {
-        heap.get_mut(arr_ref).unwrap().fields[i] = Slot::Int(c as i32);
-    }
-    Ok(Some(Slot::Reference(Some(arr_ref))))
-}
-
-// ---- Integer natives ----
-
-/// Native: `Integer.parseInt(String)` — parses string to int.
-fn native_integer_parseint(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: i32 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Int(val)))
-}
-
-/// Native: `Integer.valueOf(int)` — boxes int into Integer object.
-fn native_integer_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Integer".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Int(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Integer.intValue()` — unboxes Integer to int.
-fn native_integer_intvalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-/// Native: `Integer.toString(int)` — static, converts int to String.
-fn native_integer_tostring_static(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Integer.compareTo(Object)` — compares two boxed Integers.
-fn native_integer_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let int_val = |s: &Slot| -> VmResult<i32> {
-        match s {
-            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
-                Some(Slot::Int(n)) => Ok(*n),
-                _ => Err(VmError::InvalidRef { address: *r }),
-            },
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => int_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => int_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(Slot::Int(ordering_to_int(a.cmp(&b)))))
-}
-
-// ---- String.valueOf overloads ----
-
-/// Native: `String.valueOf(long)` — converts long to String.
-fn native_string_value_of_long(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(double)` — converts double to String.
-fn native_string_value_of_double(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(float)` — converts float to String.
-fn native_string_value_of_float(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(boolean)` — converts boolean to String.
-fn native_string_value_of_boolean(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(boolean)",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(if val { "true" } else { "false" }.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(char)` — converts char to String.
-fn native_string_value_of_char(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(char)",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(Object)` — converts Object to String.
-fn native_string_value_of_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap_object_to_string(heap.get(*r)?, *r);
-            let r = heap.allocate_string(s);
-            Ok(Some(Slot::Reference(Some(r))))
-        }
-        Some(Slot::Reference(None)) => {
-            let r = heap.allocate_string("null".to_string());
-            Ok(Some(Slot::Reference(Some(r))))
-        }
-        _ => Err(VmError::TypeMismatch {
-            expected: "Reference",
-            got: "other",
-        }),
-    }
-}
-
-// ---- String.concat ----
-
-/// Native: `String.concat(String)` — concatenates two strings.
-fn native_string_concat(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s1 = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let other_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s2 = heap
-        .get(other_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let r = heap.allocate_string(format!("{s1}{s2}"));
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Formats a single boxed slot value using the given format specifier.
-fn format_arg(
-    spec: char,
-    precision: Option<usize>,
-    slot: &Slot,
-    heap: &duke_gc::Heap,
-) -> VmResult<String> {
-    match slot {
-        Slot::Reference(None) => Ok("null".to_string()),
-        Slot::Reference(Some(r)) => {
-            let obj = heap.get(*r)?;
-            match spec {
-                's' => Ok(heap_object_to_string(obj, *r)),
-                'd' => match obj.fields.first() {
-                    Some(Slot::Int(v)) => Ok(v.to_string()),
-                    Some(Slot::Long(v)) => Ok(v.to_string()),
-                    _ => Ok("0".to_string()),
-                },
-                'f' => {
-                    let v = match obj.fields.first() {
-                        Some(Slot::Double(v)) => *v,
-                        Some(Slot::Float(v)) => f64::from(*v),
-                        _ => 0.0,
-                    };
-                    Ok(precision.map_or_else(|| format!("{v:.6}"), |p| format!("{v:.p$}")))
-                }
-                'x' => match obj.fields.first() {
-                    Some(Slot::Int(v)) => Ok(format!("{v:x}")),
-                    Some(Slot::Long(v)) => Ok(format!("{v:x}")),
-                    _ => Ok("0".to_string()),
-                },
-                'X' => match obj.fields.first() {
-                    Some(Slot::Int(v)) => Ok(format!("{v:X}")),
-                    Some(Slot::Long(v)) => Ok(format!("{v:X}")),
-                    _ => Ok("0".to_string()),
-                },
-                _ => Ok(String::new()),
-            }
-        }
-        _ => Ok(String::new()),
-    }
-}
-
-/// Native: `String.format(Ljava/lang/String;[Ljava/lang/Object;)Ljava/lang/String;`
-fn native_string_format(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let fmt_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fmt = heap.get(fmt_ref)?.string_value.clone().unwrap_or_default();
-
-    let arr_len = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap.get(*r)?.fields.len(),
-        _ => 0,
-    };
-
-    let mut result = String::new();
-    let mut arg_idx = 0usize;
-    let chars: Vec<char> = fmt.chars().collect();
-    let mut i = 0;
-
-    while i < chars.len() {
-        if chars[i] != '%' {
-            result.push(chars[i]);
-            i += 1;
-            continue;
-        }
-        i += 1;
-        if i >= chars.len() {
-            break;
-        }
-
-        // Parse optional precision: %.2f
-        let precision: Option<usize> = if chars[i] == '.' {
-            i += 1;
-            let mut prec_str = String::new();
-            while i < chars.len() && chars[i].is_ascii_digit() {
-                prec_str.push(chars[i]);
-                i += 1;
-            }
-            prec_str.parse().ok()
-        } else {
-            None
-        };
-
-        // Skip optional width digits
-        while i < chars.len() && chars[i].is_ascii_digit() {
-            i += 1;
-        }
-
-        if i >= chars.len() {
-            break;
-        }
-        let spec = chars[i];
-        i += 1;
-
-        match spec {
-            '%' => result.push('%'),
-            'n' => result.push('\n'),
-            's' | 'd' | 'f' | 'x' | 'X' => {
-                let slot = if arg_idx < arr_len {
-                    match args.get(1) {
-                        Some(Slot::Reference(Some(r))) => heap
-                            .get(*r)?
-                            .fields
-                            .get(arg_idx)
-                            .copied()
-                            .unwrap_or(Slot::Reference(None)),
-                        _ => Slot::Reference(None),
-                    }
-                } else {
-                    Slot::Reference(None)
-                };
-                arg_idx += 1;
-                let formatted = format_arg(spec, precision, &slot, heap)?;
-                result.push_str(&formatted);
-            }
-            _ => {
-                result.push('%');
-                result.push(spec);
-            }
-        }
-    }
-
-    let r = heap.allocate_string(result);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-// ---- Extended String natives ----
-
-/// Native: `String.toUpperCase()` — returns a new uppercase String.
-fn native_string_touppercase(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let r = heap.allocate_string(s.to_uppercase());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.toLowerCase()` — returns a new lowercase String.
-fn native_string_tolowercase(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let r = heap.allocate_string(s.to_lowercase());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.replace(char, char)` — replaces all occurrences of old char with new char.
-#[allow(clippy::cast_sign_loss)]
-fn native_string_replace_char(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let old_char = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let new_char = match args.get(2) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let result = s.replace(old_char, &new_char.to_string());
-    let r = heap.allocate_string(result);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.replace(CharSequence, CharSequence)` — replaces all occurrences of target with replacement.
-fn native_string_replace_charsequence(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let target_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let target = heap
-        .get(target_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let replacement_ref = match args.get(2) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let replacement = heap
-        .get(replacement_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let result = s.replace(&*target, &replacement);
-    let r = heap.allocate_string(result);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.split(String)` — splits string by delimiter, returns String array.
-fn native_string_split(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let delim_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let delim = heap
-        .get(delim_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let parts: Vec<&str> = s.split(&*delim).collect();
-    let arr_ref = heap.allocate("[Ljava/lang/String;".to_string(), parts.len());
-    for (i, part) in parts.iter().enumerate() {
-        let str_ref = heap.allocate_string((*part).to_string());
-        heap.get_mut(arr_ref).unwrap().fields[i] = Slot::Reference(Some(str_ref));
-    }
-    Ok(Some(Slot::Reference(Some(arr_ref))))
-}
-
-/// Native: `String.hashCode()` — Java's hash algorithm: `s[0]*31^(n-1) + s[1]*31^(n-2) + ... + s[n-1]`.
-#[allow(clippy::cast_possible_wrap)]
-fn native_string_hashcode(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let mut h: i32 = 0;
-    for ch in s.chars() {
-        h = h.wrapping_mul(31).wrapping_add(ch as i32);
-    }
-    Ok(Some(Slot::Int(h)))
-}
-
-/// Native: `String.toString()` — identity, returns `this`.
-#[allow(clippy::unnecessary_wraps)]
-fn native_string_tostring(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    Ok(Some(args.first().copied().unwrap_or(Slot::Reference(None))))
-}
-
-// ---- Math natives ----
-
-/// Native: `Math.max(int, int)` — returns the larger value.
-fn native_math_max_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Int(a.max(b))))
-}
-
-/// Native: `Math.min(int, int)` — returns the smaller value.
-fn native_math_min_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Int(a.min(b))))
-}
-
-/// Native: `Math.abs(int)` — returns absolute value.
-fn native_math_abs_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Int(a.wrapping_abs())))
-}
-
-// ---- Extended Math natives ----
-
-/// Native: `Math.sqrt(double)` — returns square root.
-fn native_math_sqrt(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.sqrt())))
-}
-
-/// Native: `Math.pow(double, double)` — returns a raised to the power b.
-fn native_math_pow(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.powf(b))))
-}
-
-/// Native: `Math.floor(double)` — returns floor value.
-fn native_math_floor(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.floor())))
-}
-
-/// Native: `Math.ceil(double)` — returns ceiling value.
-fn native_math_ceil(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.ceil())))
-}
-
-/// Native: `Math.round(double)` — returns closest long.
-#[allow(clippy::cast_possible_truncation)]
-fn native_math_round_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.round() as i64)))
-}
-
-/// Native: `Math.abs(long)` — returns absolute value.
-fn native_math_abs_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.wrapping_abs())))
-}
-
-/// Native: `Math.abs(double)` — returns absolute value.
-fn native_math_abs_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.abs())))
-}
-
-/// Native: `Math.max(long, long)` — returns the larger value.
-fn native_math_max_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.max(b))))
-}
-
-/// Native: `Math.min(long, long)` — returns the smaller value.
-fn native_math_min_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.min(b))))
-}
-
-/// Native: `Math.max(double, double)` — returns the larger value.
-fn native_math_max_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.max(b))))
-}
-
-/// Native: `Math.min(double, double)` — returns the smaller value.
-fn native_math_min_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.min(b))))
-}
-
-// ---- Long class natives ----
-
-/// Native: `Long.parseLong(String)` — parses string to long.
-fn native_long_parselong(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: i64 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Long(val)))
-}
-
-/// Native: `Long.valueOf(long)` — boxes long into Long object.
-fn native_long_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Long".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Long(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Long.longValue()` — unboxes Long to long.
-fn native_long_longvalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-/// Native: `Long.toString(long)` — static, converts long to String.
-fn native_long_tostring_static(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Long.compareTo(Object)` — compares two boxed Longs.
-fn native_long_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let long_val = |s: &Slot| -> VmResult<i64> {
-        match s {
-            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
-                Some(Slot::Long(n)) => Ok(*n),
-                _ => Err(VmError::InvalidRef { address: *r }),
-            },
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => long_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => long_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(Slot::Int(ordering_to_int(a.cmp(&b)))))
-}
-
-// ---- Double class natives ----
-
-/// Native: `Double.parseDouble(String)` — parses string to double.
-fn native_double_parsedouble(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: f64 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Double(val)))
-}
-
-/// Native: `Double.valueOf(double)` — boxes double into Double object.
-fn native_double_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Double".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Double(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Double.doubleValue()` — unboxes Double to double.
-fn native_double_doublevalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-// ---- Float class native ----
-
-/// Native: `Float.parseFloat(String)` — parses string to float.
-fn native_float_parsefloat(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: f32 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Float(val)))
-}
-
-// ---- Boolean class native ----
-
-/// Native: `Boolean.parseBoolean(String)` — case-insensitive "true" → 1, else 0.
-fn native_boolean_parseboolean(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap.get(*r)?.string_value.clone().unwrap_or_default();
-            let val = s.eq_ignore_ascii_case("true");
-            Ok(Some(Slot::Int(i32::from(val))))
-        }
-        Some(Slot::Reference(None)) => Ok(Some(Slot::Int(0))),
-        _ => Err(VmError::TypeMismatch {
-            expected: "Reference",
-            got: "other",
-        }),
-    }
-}
-
-/// Execute a `StringConcatFactory` recipe: walk the recipe string, replacing
-/// `\u{1}` placeholders with stringified dynamic args from the operand stack.
-fn execute_string_concat_recipe(
-    recipe: &str,
-    dynamic_args: &[Slot],
-    arg_types: &[char],
-    constants: &[String],
-    heap: &mut duke_gc::Heap,
-) -> VmResult<Slot> {
-    let mut result = String::new();
-    let mut dyn_idx = 0;
-    let mut const_idx = 0;
-
-    for ch in recipe.chars() {
-        match ch {
-            '\u{1}' => {
-                if dyn_idx < dynamic_args.len() {
-                    let type_hint = arg_types.get(dyn_idx).copied().unwrap_or('I');
-                    stringify_slot(&dynamic_args[dyn_idx], type_hint, heap, &mut result)?;
-                    dyn_idx += 1;
-                }
-            }
-            '\u{2}' => {
-                if const_idx < constants.len() {
-                    result.push_str(&constants[const_idx]);
-                    const_idx += 1;
-                }
-            }
-            other => result.push(other),
-        }
-    }
-
-    let r = heap.allocate_string(result);
-    Ok(Slot::Reference(Some(r)))
-}
-
-/// Convert a Slot to its string representation (like Java's String.valueOf).
-/// `type_hint` is the JVM type descriptor char: 'Z' for boolean, 'I' for int, etc.
-fn stringify_slot(
-    slot: &Slot,
-    type_hint: char,
-    heap: &duke_gc::Heap,
-    out: &mut String,
-) -> VmResult<()> {
-    match slot {
-        Slot::Int(v) => {
-            if type_hint == 'Z' {
-                // JVM boolean: 0 = false, nonzero = true.
-                out.push_str(if *v != 0 { "true" } else { "false" });
-            } else if type_hint == 'C' {
-                // JVM char: render as the Unicode character.
-                if let Some(ch) = char::from_u32((*v).cast_unsigned()) {
-                    out.push(ch);
-                } else {
-                    out.push('?');
-                }
-            } else {
-                out.push_str(&v.to_string());
-            }
-        }
-        Slot::Long(v) => out.push_str(&v.to_string()),
-        Slot::Float(v) => out.push_str(&format_java_float(*v)),
-        Slot::Double(v) => out.push_str(&format_java_double(*v)),
-        Slot::Reference(None) => out.push_str("null"),
-        Slot::Reference(Some(r)) => {
-            out.push_str(&heap_object_to_string(heap.get(*r)?, *r));
-        }
-        Slot::ReturnAddress(v) => out.push_str(&v.to_string()),
-    }
-    Ok(())
-}
-
-/// Format a float like Java's Float.toString.
-fn format_java_float(v: f32) -> String {
-    if v.is_nan() {
-        return "NaN".to_string();
-    }
-    if v.is_infinite() {
-        return if v > 0.0 { "Infinity" } else { "-Infinity" }.to_string();
-    }
-    let s = format!("{v}");
-    if s.contains('.') { s } else { format!("{v}.0") }
-}
-
-/// Format a double like Java's Double.toString.
-fn format_java_double(v: f64) -> String {
-    if v.is_nan() {
-        return "NaN".to_string();
-    }
-    if v.is_infinite() {
-        return if v > 0.0 { "Infinity" } else { "-Infinity" }.to_string();
-    }
-    let s = format!("{v}");
-    if s.contains('.') { s } else { format!("{v}.0") }
-}
-
-/// Execute a decoded JVM instruction stream.
-///
-/// # Parameters
-/// - `instructions`: output of `duke_bytecode::decode`
-/// - `cp`: constant pool from the parsed `ClassFile`
-/// - `args`: initial local variable values (method arguments)
-/// - `max_stack`: `Code.max_stack` from the class file
-/// - `max_locals`: `Code.max_locals` from the class file
-///
-/// # Returns
-/// `Ok(Some(slot))` for value-returning methods, `Ok(None)` for `void`.
-///
-/// # Errors
-/// Returns [`VmError`] on execution faults (division by zero, stack overflow,
-/// unimplemented instruction, etc.).
-#[allow(
-    clippy::cast_sign_loss,
-    clippy::cast_possible_truncation,
-    clippy::cast_possible_wrap,
-    clippy::cast_precision_loss,
-    clippy::too_many_lines
-)]
 pub fn execute(
     instructions: &[(usize, Instruction)],
     cp: &[Option<CpEntry>],
@@ -9213,7 +3917,7 @@ fn run_execution(
                         (recipe, consts)
                     };
 
-                    let result = execute_string_concat_recipe(
+                    let result = crate::natives::execute_string_concat_recipe(
                         &recipe,
                         &dynamic_args,
                         &arg_types,
@@ -9900,7 +4604,11 @@ fn resolve_thread_entry(
         )));
     }
 
-    let target_ref = match heap.get(thread_ref)?.fields.get(THREAD_TARGET_SLOT) {
+    let target_ref = match heap
+        .get(thread_ref)?
+        .fields
+        .get(crate::natives::THREAD_TARGET_SLOT)
+    {
         Some(Slot::Reference(Some(target_ref))) => *target_ref,
         _ => return Ok(None),
     };
@@ -10035,7 +4743,7 @@ fn spawn_java_thread(
         let mut shared_guard = shared.lock().unwrap();
         let thread = shared_guard.heap.get_mut(thread_ref)?;
         let already_started = matches!(
-            thread.fields.get(THREAD_ID_SLOT),
+            thread.fields.get(crate::natives::THREAD_ID_SLOT),
             Some(Slot::Int(thread_id)) if *thread_id >= 0
         );
         drop(shared_guard);
@@ -10057,7 +4765,7 @@ fn spawn_java_thread(
         let mut shared_guard = shared.lock().unwrap();
         {
             let thread = shared_guard.heap.get_mut(thread_ref)?;
-            thread.fields[THREAD_ID_SLOT] = Slot::Int(thread_id);
+            thread.fields[crate::natives::THREAD_ID_SLOT] = Slot::Int(thread_id);
         }
         let CompletionVm {
             registry,
@@ -11950,8 +6658,8 @@ fn array_list_sort(
                 heap,
                 output,
                 &class_name,
-                COMPARE_TO_METHOD,
-                COMPARE_TO_OBJECT_DESC,
+                crate::natives::COMPARE_TO_METHOD,
+                crate::natives::COMPARE_TO_OBJECT_DESC,
                 vec![Slot::Reference(Some(receiver)), Slot::Reference(Some(key))],
             )?;
 
@@ -12029,7 +6737,7 @@ fn native_collections_sort(
         output,
         &class_name,
         "sort",
-        SORT_COMPARATOR_DESC,
+        crate::natives::SORT_COMPARATOR_DESC,
         vec![Slot::Reference(Some(list_ref)), Slot::Reference(None)],
     )?;
     Ok(None)
@@ -12157,7 +6865,9 @@ fn native_double_compareto(
         None => return Err(VmError::NullPointerException),
     };
     // Use total_cmp: implements Java's total order where NaN > +∞ > … > -∞.
-    Ok(Some(Slot::Int(ordering_to_int(a.total_cmp(&b)))))
+    Ok(Some(Slot::Int(crate::natives::ordering_to_int(
+        a.total_cmp(&b),
+    ))))
 }
 
 // ---- Arrays natives ----
